+++ src/Maskbox.cc
@@ -101,19 +101,19 @@ int ReadCoords(double *xvals, double *yvals, const char *polyfile, FILE *fp)
 }
 
 
-void genlonlatbox(int argc_offset, int gridID1, int *lat1, int *lat2, int *lon11, int *lon12, int *lon21, int *lon22);
+void genlonlatbox(int argc_offset, int gridID1, long *lat1, long *lat2, long *lon11, long *lon12, long *lon21, long *lon22);
 
-void genindexbox(int argc_offset, int gridID1, int *lat1, int *lat2, int *lon11, int *lon12, int *lon21, int *lon22);
+void genindexbox(int argc_offset, int gridID1, long *lat1, long *lat2, long *lon11, long *lon12, long *lon21, long *lon22);
 
 
 static
-void maskbox(bool *mask, int gridID, int lat1, int lat2, int lon11, int lon12, int lon21, int lon22)
+void maskbox(bool *mask, int gridID, long lat1, long lat2, long lon11, long lon12, long lon21, long lon22)
 {
-  size_t nlon = gridInqXsize(gridID);
-  int nlat = gridInqYsize(gridID);
+  long nlon = gridInqXsize(gridID);
+  long nlat = gridInqYsize(gridID);
 
-  for ( int ilat = 0; ilat < nlat; ilat++ )
-    for ( int ilon = 0; ilon < nlon; ilon++ )
+  for ( long ilat = 0; ilat < nlat; ilat++ )
+    for ( long ilon = 0; ilon < nlon; ilon++ )
       if (  (lat1 <= ilat && ilat <= lat2 && 
 	      ((lon11 <= ilon && ilon <= lon12) || (lon21 <= ilon && ilon <= lon22))) )
 	mask[nlon*ilat + ilon] = false;
@@ -178,8 +178,8 @@ void maskbox_cell(bool *mask, int gridID)
 static
 void maskregion(bool *mask, int gridID, double *xcoords, double *ycoords, int nofcoords)
 {
-  size_t nlon = gridInqXsize(gridID);
-  int nlat = gridInqYsize(gridID);
+  long nlon = gridInqXsize(gridID);
+  long nlat = gridInqYsize(gridID);
 
   double *xvals = (double*) Malloc(nlon*sizeof(double));
   double *yvals = (double*) Malloc(nlat*sizeof(double));
@@ -201,24 +201,24 @@ void maskregion(bool *mask, int gridID, double *xcoords, double *ycoords, int no
   double ymin = yvals[0];
   double ymax = yvals[0];
 
-  for ( int i = 1; i < nlon; i++ )
+  for ( long i = 1; i < nlon; i++ )
     {
       if ( xvals[i] < xmin ) xmin = xvals[i];
       if ( xvals[i] > xmax ) xmax = xvals[i];
     }
 
-  for ( int i = 1; i < nlat; i++ )
+  for ( long i = 1; i < nlat; i++ )
     {
       if ( yvals[i] < ymin ) ymin = yvals[i];
       if ( yvals[i] > ymax ) ymax = yvals[i];
     }
 
-  for ( int ilat = 0; ilat < nlat; ilat++ )
+  for ( long ilat = 0; ilat < nlat; ilat++ )
     {
       double yval = yvals[ilat];
-      for ( int ilon = 0; ilon < nlon; ilon++ )
+      for ( long ilon = 0; ilon < nlon; ilon++ )
 	{
-          int i, j;
+          long i, j;
           int c = 0;
 	  double xval = xvals[ilon];
 	  if (!( ( ( xval > xmin ) || ( xval < xmax ) ) || ( (yval > ymin) || (yval < ymax) ) ) ) c = !c;
@@ -276,7 +276,7 @@ void *Maskbox(void *process)
   int varID, levelID;
   int gridID = -1;
   int index, gridtype = -1;
-  int lat1, lat2, lon11, lon12, lon21, lon22;
+  long lat1, lat2, lon11, lon12, lon21, lon22;
 
   cdoInitialize(process);
